fix(product): keep the launch-intent registry off the chat-surface module (PRO-230)
`launchIntentOwnsShell` is a pure predicate over a `LaunchIntentScope` and the
shell's ids — it never touches `ChatSurfaceState`. Living in `chat-surface.ts`
made the new launch-intent registry import the whole surface-arbitration module,
which the login first-load graph reaches through the session-creation actions;
that is ~2.6 kB of surface code the signed-out shell parses for nothing.

Move it (and its cases) to `launch-intent.ts`, where its inputs already live.
`chat-surface.ts` now imports it from there, which is the direction the
dependency already ran.
