+++ apps/packages/product-client/src/lib/domain/chat/launch/launch-intent.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, it } from "vitest";
 import type { ChatLaunchIntent } from "#product/lib/domain/chat/launch/launch-intent";
 import type { PendingWorkspaceEntry } from "#product/lib/domain/workspaces/creation/pending-entry";
 import {
+  launchIntentOwnsShell,
   resolveChatLaunchIntentView,
   resolveChatLaunchRetryMode,
   resolveLaunchIntentPendingAttemptId,
@@ -207,5 +208,47 @@ describe("resolveLaunchIntentScope", () => {
       pendingUiKey: null,
       workspaceId: "workspace-real",
     });
+  });
+});
+
+describe("launchIntentOwnsShell", () => {
+  it("owns an empty shell when unscoped", () => {
+    expect(launchIntentOwnsShell({
+      scope: null,
+      shellLogicalWorkspaceId: null,
+      shellWorkspaceId: null,
+    })).toBe(true);
+  });
+
+  it("does not own a selected shell when unscoped", () => {
+    expect(launchIntentOwnsShell({
+      scope: null,
+      shellLogicalWorkspaceId: "workspace-1",
+      shellWorkspaceId: "workspace-1",
+    })).toBe(false);
+  });
+
+  it("owns the shell matching its pending UI key", () => {
+    expect(launchIntentOwnsShell({
+      scope: { pendingUiKey: "pending-workspace:attempt-1", workspaceId: null },
+      shellLogicalWorkspaceId: "pending-workspace:attempt-1",
+      shellWorkspaceId: null,
+    })).toBe(true);
+  });
+
+  it("does not own an unrelated shell when scoped", () => {
+    expect(launchIntentOwnsShell({
+      scope: { pendingUiKey: null, workspaceId: "workspace-a" },
+      shellLogicalWorkspaceId: "workspace-b",
+      shellWorkspaceId: "workspace-b",
+    })).toBe(false);
+  });
+
+  it("owns the shell matching its materialized/target workspace id", () => {
+    expect(launchIntentOwnsShell({
+      scope: { pendingUiKey: null, workspaceId: "workspace-a" },
+      shellLogicalWorkspaceId: "workspace-a",
+      shellWorkspaceId: "workspace-a",
+    })).toBe(true);
   });
 });
